Initial commit
My only contribution to this repo is the text document titled “MYCUSTOMFILTER”. It is a collection of ad and tracking domains that should be but are currently not included in the three base filters for the privacy module. These domains are purely ad-blocking and privacy enhancements, and will not disrupt an app’s functionality when blocked.